Add Contacts tab and refactor
