+++ app/src/main/java/com/andela/helpmebuy/fragments/TravelListFragment.java
@@ -34,7 +34,7 @@ import com.pnikosis.materialishprogress.ProgressWheel;
 import java.util.ArrayList;
 import java.util.List;
 
-public class HomeActivityFragment extends Fragment implements CurrentTravelListener {
+public class TravelListFragment extends Fragment implements CurrentTravelListener {
 
     public final static String TAG = "HomeActivity";
 
@@ -86,8 +86,6 @@ public class HomeActivityFragment extends Fragment implements CurrentTravelListe
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
                 changeLocation();
 
             }
